from	Simon Crosby <simc@xtra.co.nz> to	mpj-user@lists.reading.ac.uk date	Wed, May 20, 2009 at 7:47 AM subject	[Mpj-user] Possible bug for mailing list	<mpj-user.lists.reading.ac.uk> Filter messages from this mailing list mailed-by	gmail.com
hide details 7:47 AM (2 hours ago)
	
	
Reply
	
	Follow up message

Hello,

I am trying to use MPJ Express to send booleans from one machine to 
another. I believe that I have found a bug. I had a quick look through 
the archives and didn't see this posted before.

There seems to be a problem where after the first "true" in the boolean 
array all the elements in the rest of the array are also set to true.