Fixed the following bug in the MPJ Express runtime code:
 The bug I (think I) found is triggered when one has a machine file with more machines than the user requests (via the -np switch on the mpjrun script). Specifically, in MPJRun::assignTasks in the branch where nprocs <= noOfMachines. The CONF_FILE_CONTENTS is being updated via code similar to:

       if(deviceName.equals("niodev")) {
          CONF_FILE_CONTENTS += ";" +name + "@"
                                    + MPJ_SERVER_PORT + "@" + (rank++) ;
        } else if(deviceName.equals("mxdev")) {
       ....

However, in the loop where this code resides, the variable `name` is not assigned, it is always == null. Hence the file that is forwarded to Wrapper is not correct (file names are all null). This causes the remote node to throw exceptions from the xdev sub-system. Something like:

Caused by: xdev.XDevException: java.nio.channels.UnresolvedAddressException
        at xdev.niodev.NIODevice.init(NIODevice.java:817)
        at mpjdev.MPJDev.init(MPJDev.java:89)
        at mpi.MPI.Init(MPI.java:253)